requirements.txt
alembic
bidict
cachelib
click
cryptography
email_validator
Flask
Flask-Cors
Flask-Mail
Flask-Migrate
Flask-Session
Flask-SQLAlchemy
Flask-WTF
Pillow
PyJWT
SQLAlchemy
Werkzeug
WTForms
Markup
